updated Content tests
